Move LaunchGame into Game.ts
