+++ src/renderer/pages/LauncherPage.tsx
@@ -1,91 +1,23 @@
 import Dropdown from "../components/Dropdown";
 import MinecraftButton from "../components/MinecraftButton";
 import {useAppState} from "../contexts/AppState";
-import {SemVersion} from "../scripts/classes/SemVersion";
-import { GetLauncherConfig, SetLauncherConfig } from "../scripts/Launcher";
-import { IsDevModeEnabled, TryEnableDevMode } from "../scripts/DeveloperMode";
-import { RegisterVersion, UnregisterCurrent } from "../scripts/AppRegistry";
-import { CleanupInstall, InstallProxy, CreateLock, DownloadVersion, ExtractVersion, IsLocked, IsRegistered, IsDownloaded } from "../scripts/VersionManager";
 
-import * as child from 'child_process';
+import {LaunchGame} from "../scripts/Game";
 
 export default function LauncherPage() {
     const {
-        allProfiles, selectedProfile, setSelectedProfile, loadingPercent, status, setStatus, isLoading, setIsLoading,
-        setLoadingPercent, allMinecraftVersions, error, setError
+        allProfiles,
+        selectedProfile, setSelectedProfile,
+        loadingPercent,
+        status, setStatus,
+        isLoading, setIsLoading,
+        error, setError
     } = useAppState();
 
-    const log = (msg: string) => {
-        console.log(msg)
-        setStatus(msg)
-    }
 
     const launchGame = async () => {
-        if (isLoading) return;
-
         try {
-            if (allProfiles.length === 0) {
-                throw new Error("Cannot launch without a profile!")
-            }
-
-            const profile = allProfiles[selectedProfile];
-            const semVersion = SemVersion.fromString(profile.minecraft_version);
-            const minecraftVersion = allMinecraftVersions.find(version => version.version.toString() === semVersion.toString())!;
-
-            if (minecraftVersion === undefined) {
-                throw new Error(`Failed to find minecraft version ${semVersion.toString()} in the profile in allVersions!`);
-            }
-
-            setError("");
-            setIsLoading(true);
-
-            // Check that the user has developer mode enabled on windows for the game to be installed through loose files.
-            if (!IsDevModeEnabled()) {
-                const enabled_dev = await TryEnableDevMode();
-                if (!enabled_dev) {
-                    throw new Error("Failed to enable 'Developer Mode' in windows settings to allow installing the game from loose files, please enable manually or make sure to press 'Yes' to enable automatically.")
-                }
-            }
-
-            // We create a lock file when starting the download
-            // if we are doing a launch, and we detect it for the version we are targeting
-            // there is a good chance the previous install/download failed and therefore remove it.
-            const didPreviousDownloadFail = IsLocked(semVersion);
-            
-            if (didPreviousDownloadFail) {
-                log("Detected a .lock file from the previous download attempt, cleaning up.");
-                CleanupInstall(semVersion, false);
-                log("Removed previous download attempt.");
-            }
-
-            // Check for the folder for the version we are targeting, if not present we need to fetch.
-            if (!IsDownloaded(semVersion)) {
-                log("Target version is not downloaded.");
-                CreateLock(semVersion);
-                await DownloadVersion(minecraftVersion, setStatus, setLoadingPercent);
-                await ExtractVersion(minecraftVersion, setStatus, setLoadingPercent);
-                log("Cleaning up after successful download")
-                CleanupInstall(semVersion, true);
-            }
-
-            // Only register the game if needed
-            if (!IsRegistered(minecraftVersion)) {
-                setStatus("Unregistering existing version");
-                await UnregisterCurrent();
-
-                setStatus("Registering downloaded version");
-                await RegisterVersion(minecraftVersion)
-
-                SetLauncherConfig(GetLauncherConfig());
-            }
-
-            setIsLoading(false);
-            setStatus("");
-
-            InstallProxy(minecraftVersion);
-
-            const startGameCmd = `start minecraft:`;
-            child.exec(startGameCmd)
+            await LaunchGame()
         } catch (e) {
             console.error(e);
             setError((e as Error).message);
